Debounce buttons by setting a time interval after pushing a button.
When button is pushed there is an interrupt called which starts a timer.
The timer counts 50ms and calls an interrupt routine which checks the
last button pushed (last_button_pushed variable) and execute the routine
that describes this button behaviour.
The router behaviour function checks again if button is set and if so
performs its body. Theoretically this should prevent buttons from  bouncing.